requirements.txt
openai
flask[async]
python-dotenv
langchain
pypdf
faiss-cpu
tiktoken
asyncio
futures
werkzeug
pinecone-client
tree-of-thoughts
torch
Flask-Dance
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
flask_login
httplib2
